trainer/requirements.txt
acme==0.31.0
asn1crypto==0.24.0
binwalk==2.1.1
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2020.6.20
chardet==3.0.4
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==2.1.4
cycler==0.10.0
decorator==4.1.2
distro-info===0.18ubuntu0.18.04.1
dnspython==1.16.0
future==0.15.2
idna==2.9
joblib==0.15.1
josepy==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
matplotlib==2.1.1
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
numpy==1.19.0
olefile==0.45.1
pandas==1.0.5
parsedatetime==2.4
pbr==3.1.1
Pillow==5.1.0
psutil==5.6.7
pyasn1==0.4.2
pycrypto==2.6.1
pydig==0.3.0
pygobject==3.26.1
PyICU==1.9.8
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyqtgraph==0.10.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-augeas==0.5.0
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.8.0
scikit-learn==0.23.1
scipy==1.5.0
SecretStorage==2.3.1
six==1.15.0
sklearn==0.0
sklearn-rvm==0.1.1
ssh-import-id==5.7
threadpoolctl==2.1.0
tld==0.12.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
vboxapi==1.0
whois==0.9.7
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
